+++ SpicesEngine/src/Render/Renderer/SpecificRenderer/RayTracingRenderer.cpp
@@ -244,8 +244,8 @@ namespace Spices {
 
 		const auto rayTracingMaterial                 = ResourcePool<Material>::Load<Material>("RayTracingRenderer.RayTracing.Default");
 
-		const uint32_t rayGenCount              = rayTracingMaterial->GetShaderPath("rgen").size();
-		const uint32_t missCount                = rayTracingMaterial->GetShaderPath("rmiss").size();
+		const uint32_t rayGenCount              = static_cast<uint32_t>(rayTracingMaterial->GetShaderPath("rgen").size());
+		const uint32_t missCount                = static_cast<uint32_t>(rayTracingMaterial->GetShaderPath("rmiss").size());
 		const uint32_t hitCount                 = static_cast<uint32_t>(m_HitGroups.size());
 
 		const auto handleCount          = rayGenCount + missCount + hitCount;
